[fix]: disable transitions on theme change
